Addressing feedback related to instruction spec table
